[YUNIKORN-3057] Refactor basic_scheduling E2E test suite structure
